requirements.txt
discord.py
pynacl
python-dotenv
google-genai